new BRANCH == dev.collin/BROKEN/no-sticky-footer/REPAIRS-UNDERWAY/180825-0
F: sticky footer == BROKEN
.................................
renamed the faculty.html page to guitarFaculty.html
.................................
forced checkpoint as I am renaming stuff and looking for what is not working and what is not used
at this point no further regressions...

----------------------------------
when I use bg colors, the middle content section colors leave short verticle bars at the top left and bottom left
and the footer color ends before any of the bottons appear. So I dont understand this and forcing the current code
to reside on this branch, before immediately branching again to fix this problem.